Added rectangular filter and made masks content-aware
